requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
certifi==2020.12.5
Django==3.1.4
django-graphql-playground==1.0
django-jet==1.0.8
djangorestframework==3.12.2
graphene==2.1.8
graphene-django==2.14.0
graphene-generator==0.1.2
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==8.0.1
promise==2.3
pytz==2020.5
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
wincertstore==0.2
